node: determine roamability when selecting neighbors
Currently, the neighbor report list is sorted based on the total amount
of roam actions to and from the specific foreign nodes.

This does not take into account that nodes might not be created at the
same time. When a popular neighbor node reboots, it's roam actions are
reset and it's ranking might not recover for a long amount of time
depending on how long the local node is already running.

Change this shotcoming to take into account the timestamp at which a
node was created. This way, the ranking depends on the amount of roam
actions relative to the nodes uptime / seen-time.

Signed-off-by: David Bauer <mail@david-bauer.net>